Changed storehistory.html
Changed storehistory.html to show a table of all the orders by default; if a user is logged in, it'll show all their orders, and when a staff member is logged in, it'll show all, orders. Also added preventative code to stop a user that isn't logged in from viewing anything.

Now all that's left to do is implement a  way to reset the store filter to nothing.